Replicate string.strip() using regex
If passed a single string, the function uses regular expressions to find
whitespace on the ends and removes it. This is the same as
string.strip()

If passed an optional second argument, the function removes every
instance of it from the string. This isn't how string.strip() functions,
I just misread the problem statement from the book.